Still working on DFA for pldi2019cnn problems with type/rank determination
I am encountering a long-standing problem with dfaUpDown,
propagating constant values into sub-functions. The problem,
which apparently crops up, e.g, in UTTHorn, is along the following
lines:

We do something like 4 thorn ⍳n in one place, and 6 thorn ⍳n in another.
We have to clone "thorn" (or inline it, which amounts to the same thing),
or else we have two choices:

  - do not propagate constant values into thorn. This approach
    would cause dfa to fail, as we can't determine types/ranks
    that depend on ⍺ in ⍺⍕⍵.

  - blindly propagate constant values into thorn. This would overwrite
    one set of constant ⍺ by another, letting DFA proceed, but
    would give wrong answers for the first ⍺ call.
    Wrong answers are not wanted.

There is also a problem that I hope to fix today, of having differing
values between astvalue and astPred[astPredValue]. I don't recall
how we ended up having two instances of the same thing, but I am
going to kill the latter, as it needs killin'.